Api with validation

+++ controllers/todo.js
@@ -1,4 +1,5 @@
 const Todo = require("../models/todo");
+const { todoSchema } = require("../middleware/joi");
 
 exports.findData = async function (req, res) {
   try {
@@ -15,13 +16,19 @@ exports.findData = async function (req, res) {
 
 exports.findUser = async function (req, res) {
   try {
+    const data = await findById(req.params.id);
+    if (!data) throw new Error("id is not found");
+    const findOne = await findOne({ Title: req.body.Title });
+    if (!findOne) throw new Error("Title is not found");
     const user = await Todo.findById(req.params.id);
     console.log(user, "users");
-    res.json({
-      success: true,
-      message: "get details Successfully",
-      data: user,
-    });
+    if (!user) throw new Error("data is not found");
+    else
+      return res.json({
+        success: true,
+        message: "get details Successfully",
+        data: user,
+      });
   } catch (err) {
     res.json({ success: false, message: err.message });
   }
@@ -29,7 +36,9 @@ exports.findUser = async function (req, res) {
 
 exports.addData = async function (req, res) {
   try {
-    let createData = new Todo(req.body);
+    let value = await todoSchema.validateAsync(req.body);
+    console.log("error", value);
+    let createData = new Todo(value);
     console.log("createData", createData);
     createData.save();
     res.json({
@@ -45,7 +54,8 @@ exports.updateData = async function (req, res) {
   try {
     const data = await Todo.findByIdAndUpdate(req.params.id, req.body);
     console.log(data);
-    res.json({ success: true, message: "todo data is update", data: data });
+    if (!data) throw new Error("id is not update");
+    else res.json({ success: true, message: "todo data is update", data });
   } catch (err) {
     res.json(err);
     res.json({ success: false, message: err.message });
@@ -53,11 +63,14 @@ exports.updateData = async function (req, res) {
 };
 exports.deleteData = async function (req, res) {
   try {
+    const data = await Todo.findById(req.params.id);
+    console.log(data, "data");
+    if (!data) throw new Error("id is not found");
     const id = await Todo.findByIdAndDelete(req.params.id);
     console.log(id, "user");
     res
       .status(200)
-      .json({ success: true, message: "id delete id success", data: id });
+      .json({ success: true, message: "id delete id success", id });
   } catch (err) {
     res.json({ success: false, message: err.message });
   }
